add email templater
