+++ engine/apps/alerts/incident_appearance/templaters/__init__.py
@@ -1,6 +1,5 @@
 from .alert_templater import TemplateLoader  # noqa: F401
 from .classic_markdown_templater import AlertClassicMarkdownTemplater  # noqa: F401
-from .email_templater import AlertEmailTemplater  # noqa: F401
 from .phone_call_templater import AlertPhoneCallTemplater  # noqa: F401
 from .slack_templater import AlertSlackTemplater  # noqa: F401
 from .sms_templater import AlertSmsTemplater  # noqa: F401
