+++ engine/apps/email/backend.py
@@ -1,13 +1,14 @@
 from django.conf import settings
 from django.core.mail import send_mail
+from django.utils.html import strip_tags
 
-from apps.alerts.incident_appearance.templaters import AlertEmailTemplater
 from apps.base.messaging import BaseMessagingBackend
+from apps.email.alert_rendering import build_subject_and_title
 
 
 class EmailBackend(BaseMessagingBackend):
     backend_id = "EMAIL"
-    label = "Email 📧"
+    label = "Email"
     short_label = "Email"
     available_for_use = True
 
@@ -15,14 +16,11 @@ class EmailBackend(BaseMessagingBackend):
         return {"email": user.email}
 
     def notify_user(self, user, alert_group, notification_policy):
-        alert = alert_group.alerts.last()
-        # use existing Email templater
-        templated_alert = AlertEmailTemplater(alert).render()
+        subject, html_message = build_subject_and_title(alert_group)
 
-        title = templated_alert.title
-        message = templated_alert.message
+        message = strip_tags(html_message)
         email_from = settings.EMAIL_HOST_USER
         recipient_list = [user.email]
 
         # send email through Django setup
-        send_mail(title, message, email_from, recipient_list)
+        send_mail(subject, message, email_from, recipient_list, html_message=html_message)
